+++ src/SleetLib/FileSystem/AmazonS3FileSystem.cs
@@ -5,6 +5,8 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Amazon.S3;
+using Amazon.S3.Model;
+using Newtonsoft.Json.Linq;
 using NuGet.Common;
 using static Sleet.AmazonS3FileSystemAbstraction;
 
@@ -42,7 +44,7 @@ namespace Sleet
         {
             log.LogInformation($"Verifying {bucketName} exists.");
 
-            var isBucketFound = await client.DoesS3BucketExistAsync(bucketName).ConfigureAwait(false);
+            var isBucketFound = await HasBucket(log, token);
             if (!isBucketFound)
             {
                 log.LogError(
@@ -89,6 +91,48 @@ namespace Sleet
             }
 
             return relativePath;
+        }
+
+        public override async Task<bool> HasBucket(ILogger log, CancellationToken token)
+        {
+            return await client.DoesS3BucketExistAsync(bucketName);
+        }
+
+        public override async Task CreateBucket(ILogger log, CancellationToken token)
+        {
+            await client.EnsureBucketExistsAsync(bucketName);
+
+            var policyRequest = new PutBucketPolicyRequest()
+            {
+                BucketName = bucketName,
+                Policy = GetReadOnlyPolicy(bucketName).ToString()
+            };
+
+            await client.PutBucketPolicyAsync(policyRequest);
+        }
+
+        public override async Task DeleteBucket(ILogger log, CancellationToken token)
+        {
+            if (await HasBucket(log, token))
+            {
+                await client.DeleteBucketAsync(bucketName, token);
+            }
+        }
+
+        private static JObject GetReadOnlyPolicy(string bucketName)
+        {
+            // Grant read-only access based on https://aws.amazon.com/premiumsupport/knowledge-center/read-access-objects-s3-bucket/
+            return new JObject(
+                new JProperty("Version", "2012-10-17"),
+                new JProperty("Statement",
+                    new JArray(
+                        new JObject(
+                            new JProperty("Sid", "AllowPublicRead"),
+                            new JProperty("Effect", "Allow"),
+                            new JProperty("Principal", "*"),
+                            new JProperty("Action", new JArray("s3:GetObject")),
+                            new JProperty("Resource", new JArray($"arn:aws:s3:::{bucketName}/*"))
+                        ))));
         }
     }
 }